Split waveform tests by concern
