+++ src/gui_app/waveform.rs
@@ -262,1598 +262,4 @@ mod selection_paint;
 pub(super) type WaveformViewport = ui::IndexViewport;
 
 #[cfg(test)]
-mod tests {
-    use super::{
-        BAND_COUNT, WaveformActiveDragKind, WaveformEditFadeHandle, WaveformInteraction,
-        WaveformSelectionEdge, WaveformSelectionKind, WaveformSignalWidget, WaveformState,
-        WaveformWidget, WaveformWidgetProps, split_frequency_bands,
-        waveform_file_from_mono_samples,
-    };
-    use radiant::{
-        gui::types::{Point, Rect, Vector2},
-        runtime::{GpuSurfaceContent, PaintFillRect, PaintPrimitive},
-        theme::ThemeTokens,
-        widgets::{PointerButton, Widget, WidgetInput},
-    };
-    use std::{fs, sync::Arc};
-
-    #[test]
-    fn waveform_summary_preserves_raw_transient_detail() {
-        let samples = vec![0.0, 0.12, -0.9, 0.08, 0.0, 0.42, -0.18, 0.0];
-
-        let file = waveform_file_from_mono_samples(
-            "test.wav".into(),
-            Arc::from([]),
-            48_000,
-            1,
-            samples.clone(),
-        );
-
-        assert_eq!(BAND_COUNT, 4);
-        let raw_peak_index = samples
-            .iter()
-            .enumerate()
-            .max_by(|(_, left), (_, right)| left.abs().total_cmp(&right.abs()))
-            .map(|(index, _)| index)
-            .expect("peak sample");
-        let rendered_peak_index = file.gpu_signal_summary.levels[0]
-            .buckets
-            .iter()
-            .enumerate()
-            .max_by(|(_, left), (_, right)| {
-                left.max
-                    .abs()
-                    .max(left.min.abs())
-                    .total_cmp(&right.max.abs().max(right.min.abs()))
-            })
-            .map(|(index, _)| index / BAND_COUNT)
-            .expect("peak band sample");
-
-        assert_eq!(rendered_peak_index, raw_peak_index);
-        let frame_peak = file.gpu_signal_summary.levels[0].buckets
-            [raw_peak_index * BAND_COUNT..(raw_peak_index + 1) * BAND_COUNT]
-            .iter()
-            .map(|bucket| bucket.min.abs().max(bucket.max.abs()))
-            .fold(0.0_f32, f32::max);
-        assert!(frame_peak > 0.89);
-    }
-
-    #[test]
-    fn stereo_downmix_preserves_per_frame_peak_height_for_normalized_files() {
-        let samples = vec![1.0, 0.0, -0.25, 0.25, 0.0, -1.0, 0.5, -0.75];
-
-        assert_eq!(
-            super::downmix_to_mono(&samples, 2, 4),
-            vec![1.0, -0.25, -1.0, -0.75]
-        );
-    }
-
-    #[test]
-    fn stereo_downmix_avoids_phase_cancellation_in_visual_projection() {
-        let samples = vec![1.0, -1.0, 0.35, -0.2];
-
-        assert_eq!(super::downmix_to_mono(&samples, 2, 2), vec![1.0, 0.35]);
-    }
-
-    #[test]
-    fn playmark_extraction_writes_sibling_wav_range() {
-        let root = std::env::temp_dir().join(format!(
-            "wavecrate-playmark-extract-{}",
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .expect("clock")
-                .as_nanos()
-        ));
-        fs::create_dir_all(&root).expect("create temp root");
-        let source = root.join("source.wav");
-        write_test_wav_i16(&source, &[0, 100, 200, 300, 400, 500]);
-        let mut state = WaveformState::load_path(source).expect("load source");
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.25, 0.75));
-
-        let output = state
-            .extract_play_selection_to_sibling()
-            .expect("extract range");
-
-        assert_eq!(output.file_name().unwrap(), "source_extraction.wav");
-        assert_eq!(read_test_wav_i16(&output), vec![100, 200, 300, 400]);
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn playmark_extraction_uses_channel_independent_frame_bounds() {
-        let root = std::env::temp_dir().join(format!(
-            "wavecrate-playmark-extract-stereo-{}",
-            std::time::SystemTime::now()
-                .duration_since(std::time::UNIX_EPOCH)
-                .expect("clock")
-                .as_nanos()
-        ));
-        fs::create_dir_all(&root).expect("create temp root");
-        let source = root.join("source.wav");
-        write_test_wav_i16_stereo(
-            &source,
-            &[
-                (0, 1),
-                (100, 101),
-                (200, 201),
-                (300, 301),
-                (400, 401),
-                (500, 501),
-            ],
-        );
-        let mut state = WaveformState::load_path(source).expect("load source");
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.25, 0.75));
-
-        let output = state
-            .extract_play_selection_to_sibling()
-            .expect("extract range");
-
-        assert_eq!(
-            read_test_wav_i16(&output),
-            vec![100, 101, 200, 201, 300, 301, 400, 401]
-        );
-        let _ = fs::remove_dir_all(root);
-    }
-
-    #[test]
-    fn empty_waveform_rejects_playmark_extraction() {
-        let mut state = WaveformState::empty();
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.1, 0.2));
-
-        assert_eq!(
-            state.extract_play_selection_to_sibling(),
-            Err(String::from("Load a sample before extracting"))
-        );
-    }
-
-    #[test]
-    fn frequency_bands_keep_low_mid_high_and_raw_lanes_separate() {
-        let samples = [0.0, 0.7, -0.7, 0.18, -0.18, 0.02, -0.02, 0.0];
-        let bands = split_frequency_bands(&samples, 48_000);
-
-        assert_eq!(bands.len(), samples.len() * BAND_COUNT);
-        let low_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[0].abs())
-            .fold(0.0_f32, f32::max);
-        let mid_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[1].abs())
-            .fold(0.0_f32, f32::max);
-        let high_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[2].abs())
-            .fold(0.0_f32, f32::max);
-        let raw_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[3].abs())
-            .fold(0.0_f32, f32::max);
-
-        assert!(low_peak > 0.0);
-        assert!(mid_peak > 0.0);
-        assert!(high_peak > 0.0);
-        assert!(raw_peak > 0.69);
-    }
-
-    #[test]
-    fn frequency_bands_raw_lane_preserves_visual_peak_for_normalized_content() {
-        let sample_rate = 48_000;
-        let low = (0..sample_rate / 100)
-            .map(|frame| {
-                let t = frame as f32 / sample_rate as f32;
-                (std::f32::consts::TAU * 70.0 * t).sin()
-            })
-            .collect::<Vec<_>>();
-        let high = (0..sample_rate / 100)
-            .map(|frame| {
-                let t = frame as f32 / sample_rate as f32;
-                (std::f32::consts::TAU * 4_000.0 * t).sin()
-            })
-            .collect::<Vec<_>>();
-
-        for samples in [low, high] {
-            let bands = split_frequency_bands(&samples, sample_rate);
-            let raw_peak = bands
-                .chunks_exact(BAND_COUNT)
-                .map(|frame| frame[3].abs())
-                .fold(0.0_f32, f32::max);
-
-            assert!(
-                (raw_peak - 1.0).abs() < 0.000_01,
-                "raw display peak should track normalized sample peak, got {raw_peak}"
-            );
-        }
-    }
-
-    #[test]
-    fn frequency_bands_normalize_short_low_content_to_raw_visual_peak() {
-        let sample_rate = 48_000;
-        let samples = (0..2_656)
-            .map(|frame| {
-                let t = frame as f32 / sample_rate as f32;
-                (std::f32::consts::TAU * 72.0 * t).sin()
-            })
-            .collect::<Vec<_>>();
-
-        let bands = split_frequency_bands(&samples, sample_rate);
-        let low_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[0].abs())
-            .fold(0.0_f32, f32::max);
-        let raw_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[3].abs())
-            .fold(0.0_f32, f32::max);
-
-        assert!(raw_peak > 0.99, "raw peak was {raw_peak}");
-        assert!(
-            low_peak > raw_peak * 0.94,
-            "short low content should not render visually undersized: low={low_peak}, raw={raw_peak}"
-        );
-    }
-
-    #[test]
-    fn frequency_bands_use_envelopes_to_avoid_low_zero_crossing_gaps() {
-        let sample_rate = 48_000;
-        let samples = (0..sample_rate / 20)
-            .map(|frame| {
-                let t = frame as f32 / sample_rate as f32;
-                (std::f32::consts::TAU * 60.0 * t).sin()
-            })
-            .collect::<Vec<_>>();
-
-        let bands = split_frequency_bands(&samples, sample_rate);
-        let low_values = bands
-            .chunks_exact(BAND_COUNT)
-            .skip(sample_rate as usize / 50)
-            .map(|frame| frame[0].abs())
-            .collect::<Vec<_>>();
-        let low_peak = low_values.iter().copied().fold(0.0_f32, f32::max);
-        let low_floor = low_values.iter().copied().fold(f32::INFINITY, f32::min);
-
-        assert!(low_peak > 0.94, "low envelope peak was {low_peak}");
-        assert!(
-            low_floor > low_peak * 0.55,
-            "sustained low envelope should not collapse at zero crossings: floor={low_floor}, peak={low_peak}"
-        );
-    }
-
-    #[test]
-    fn frequency_bands_do_not_inflate_low_color_for_high_frequency_content() {
-        let sample_rate = 48_000;
-        let samples = (0..sample_rate / 100)
-            .map(|frame| {
-                let t = frame as f32 / sample_rate as f32;
-                (std::f32::consts::TAU * 7_200.0 * t).sin()
-            })
-            .collect::<Vec<_>>();
-
-        let bands = split_frequency_bands(&samples, sample_rate);
-        let low_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[0].abs())
-            .fold(0.0_f32, f32::max);
-        let high_peak = bands
-            .chunks_exact(BAND_COUNT)
-            .map(|frame| frame[2].abs())
-            .fold(0.0_f32, f32::max);
-
-        assert!(high_peak > 0.30, "high peak was {high_peak}");
-        assert!(
-            low_peak < high_peak * 0.35,
-            "mostly high-frequency content should not be painted as low-end blue: low={low_peak}, high={high_peak}"
-        );
-    }
-
-    #[test]
-    fn playback_state_starts_at_head_and_clears_on_stop() {
-        let mut state = WaveformState::synthetic_for_tests();
-
-        assert!(!state.is_playing());
-        assert_eq!(state.playhead_ratio(), None);
-        assert_eq!(state.play_mark_ratio(), None);
-
-        state.start_playback(0.0);
-        assert!(state.is_playing());
-        assert_eq!(state.playhead_ratio(), Some(0.0));
-        assert_eq!(state.play_mark_ratio(), Some(0.0));
-
-        state.set_playhead_ratio(0.375);
-        assert_eq!(state.playhead_ratio(), Some(0.375));
-        assert_eq!(state.play_mark_ratio(), Some(0.0));
-
-        state.stop_playback();
-        assert!(!state.is_playing());
-        assert_eq!(state.playhead_ratio(), None);
-        assert_eq!(state.play_mark_ratio(), Some(0.0));
-    }
-
-    #[test]
-    fn overlay_paint_projects_play_edit_and_playhead_markers() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.start_playback(0.125);
-        state.set_playhead_ratio(0.25);
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Edit,
-            visible_ratio: 0.375,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.375,
-        });
-
-        let widget = waveform_widget_for_state(&state);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(400.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let fills = fill_rects(&primitives);
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x / 400.0 - 0.125).abs() < 0.01
-                && (fill.color.r, fill.color.g, fill.color.b) == (255, 142, 92)
-                && fill.color.a == 230
-        }));
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x / 400.0 - 0.375).abs() < 0.01
-                && (fill.color.r, fill.color.g, fill.color.b) == (82, 168, 255)
-                && fill.color.a == 230
-        }));
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x / 400.0 - 0.25).abs() < 0.01
-                && (fill.color.r, fill.color.g, fill.color.b) == (71, 220, 255)
-                && fill.color.a == 245
-        }));
-    }
-
-    #[test]
-    fn playhead_cursor_paints_pixel_stable_rect_when_progress_is_subpixel() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.start_playback(0.0);
-        state.set_playhead_ratio(0.12345);
-        let widget = waveform_widget_for_state(&state);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(400.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let playhead = fill_rects(&primitives)
-            .into_iter()
-            .find(|fill| {
-                (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (71, 220, 255, 245)
-            })
-            .expect("playhead fill paints");
-        assert_eq!(playhead.rect.width(), 2.0);
-        assert_eq!(playhead.rect.min.x.fract(), 0.0);
-        assert_eq!(playhead.rect.max.x.fract(), 0.0);
-    }
-
-    #[test]
-    fn visible_ratio_maps_to_absolute_audio_position_inside_viewport() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.viewport = super::WaveformViewport {
-            start: 12_000,
-            end: 36_000,
-        };
-
-        let ratio = state.absolute_ratio_from_visible(0.5);
-
-        assert!((ratio - 0.5).abs() < 0.0001);
-    }
-
-    #[test]
-    fn auxiliary_drag_pans_zoomed_waveform_viewport() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.viewport = super::WaveformViewport {
-            start: 12_000,
-            end: 36_000,
-        };
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(100.0, 40.0),
-                    button: PointerButton::Auxiliary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("middle press should arm waveform pan");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform pan interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginPan { visible_ratio: 0.5 }
-        );
-        state.apply_interaction(interaction);
-        state.apply_interaction(WaveformInteraction::UpdateSelection {
-            visible_ratio: 0.25,
-        });
-
-        assert!(
-            state.viewport().start > 12_000,
-            "dragging left should pan the viewport later in the sample"
-        );
-        assert_eq!(state.viewport().visible_items(), 24_000);
-    }
-
-    #[test]
-    fn primary_press_emits_playback_ratio_matching_hover_cursor_ratio() {
-        let state = WaveformState::synthetic_for_tests();
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(10.0, 20.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(60.0, 40.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("playback interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginSelection {
-                kind: WaveformSelectionKind::Play,
-                visible_ratio: 0.25
-            }
-        );
-    }
-
-    #[test]
-    fn secondary_press_emits_edit_selection_begin_ratio() {
-        let state = WaveformState::synthetic_for_tests();
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(10.0, 20.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(160.0, 40.0),
-                    button: PointerButton::Secondary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("edit selection interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginSelection {
-                kind: WaveformSelectionKind::Edit,
-                visible_ratio: 0.75
-            }
-        );
-    }
-
-    #[test]
-    fn dragging_primary_creates_playmark_selection_without_starting_playback() {
-        let mut state = WaveformState::synthetic_for_tests();
-
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Play,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.6 });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.6 });
-
-        let selection = state.play_selection().expect("playmark selection");
-        assert!(!state.is_playing());
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert_eq!(state.play_mark_ratio(), Some(0.2));
-    }
-
-    #[test]
-    fn playmark_range_edges_are_resizable() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        state.play_mark_ratio = Some(0.2);
-
-        state.apply_interaction(WaveformInteraction::BeginSelectionResize {
-            kind: WaveformSelectionKind::Play,
-            edge: WaveformSelectionEdge::End,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.75,
-        });
-
-        let selection = state.play_selection().expect("playmark selection");
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.75).abs() < 0.001);
-        assert_eq!(state.play_mark_ratio(), Some(selection.start()));
-        assert!(!state.is_playing());
-    }
-
-    #[test]
-    fn playmark_top_handle_moves_range_without_resizing() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        state.play_mark_ratio = Some(0.2);
-
-        state.apply_interaction(WaveformInteraction::BeginSelectionMove {
-            kind: WaveformSelectionKind::Play,
-            visible_ratio: 0.4,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection {
-            visible_ratio: 0.55,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.55,
-        });
-
-        let selection = state.play_selection().expect("moved playmark selection");
-        assert!((selection.start() - 0.35).abs() < 0.001);
-        assert!((selection.end() - 0.75).abs() < 0.001);
-        assert!((selection.width() - 0.4).abs() < 0.001);
-        assert_eq!(state.play_mark_ratio(), Some(selection.start()));
-        assert!(!state.is_playing());
-    }
-
-    #[test]
-    fn edit_top_handle_moves_range_and_preserves_edit_effects() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_out(0.25, 0.7),
-        );
-        state.edit_mark_ratio = Some(0.2);
-
-        state.apply_interaction(WaveformInteraction::BeginSelectionMove {
-            kind: WaveformSelectionKind::Edit,
-            visible_ratio: 0.4,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.1 });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.1 });
-
-        let selection = state.edit_selection().expect("moved edit selection");
-        assert!((selection.start() - 0.0).abs() < 0.001);
-        assert!((selection.end() - 0.4).abs() < 0.001);
-        assert_eq!(state.edit_mark_ratio(), Some(selection.start()));
-        assert_eq!(selection.fade_in().map(|fade| fade.length), Some(0.25));
-        assert_eq!(selection.fade_out().map(|fade| fade.length), Some(0.25));
-    }
-
-    #[test]
-    fn primary_press_on_playmark_handle_starts_resize_instead_of_new_selection() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        state.play_mark_ratio = Some(0.2);
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(120.0, 8.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("playmark resize interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginSelectionResize {
-                kind: WaveformSelectionKind::Play,
-                edge: WaveformSelectionEdge::End,
-                visible_ratio: 0.6
-            }
-        );
-    }
-
-    #[test]
-    fn primary_press_on_playmark_top_handle_starts_move() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.play_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        state.play_mark_ratio = Some(0.2);
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(80.0, 3.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("playmark move interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginSelectionMove {
-                kind: WaveformSelectionKind::Play,
-                visible_ratio: 0.4
-            }
-        );
-    }
-
-    #[test]
-    fn secondary_press_on_edit_top_handle_starts_move() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        state.edit_mark_ratio = Some(0.2);
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(80.0, 3.0),
-                    button: PointerButton::Secondary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("edit move interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginSelectionMove {
-                kind: WaveformSelectionKind::Edit,
-                visible_ratio: 0.4
-            }
-        );
-    }
-
-    #[test]
-    fn dragging_secondary_creates_edit_selection() {
-        let mut state = WaveformState::synthetic_for_tests();
-
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Edit,
-            visible_ratio: 0.7,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection {
-            visible_ratio: 0.25,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.25,
-        });
-
-        let selection = state.edit_selection().expect("edit selection");
-        assert!((selection.start() - 0.25).abs() < 0.001);
-        assert!((selection.end() - 0.7).abs() < 0.001);
-        assert_eq!(state.edit_mark_ratio(), Some(0.7));
-    }
-
-    #[test]
-    fn edit_fade_top_handle_drag_sets_fade_in_length() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Edit,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.6 });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.6 });
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInEnd,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.3 });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.3 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade = selection.fade_in().expect("fade-in after handle drag");
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert!((fade.length - 0.25).abs() < 0.001);
-        assert!((fade.curve - 0.5).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_top_handles_push_and_restore_opposite_fade() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_out(0.25, 0.7),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInEnd,
-            visible_ratio: 0.3,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.6 });
-        let pushed = state.edit_selection().expect("pushed edit selection");
-        let pushed_fade_in = pushed.fade_in().expect("fade-in after push");
-        assert!(pushed.fade_out().is_none());
-        assert!((pushed.start() + pushed.width() * pushed_fade_in.length - 0.6).abs() < 0.001);
-
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.3 });
-        let restored = state.edit_selection().expect("restored edit selection");
-        let restored_fade_in = restored.fade_in().expect("restored fade-in");
-        let restored_fade_out = restored.fade_out().expect("restored fade-out");
-        let fade_in_end = restored.start() + restored.width() * restored_fade_in.length;
-        let fade_out_start = restored.end() - restored.width() * restored_fade_out.length;
-        assert!((fade_in_end - 0.3).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((restored_fade_in.curve - 0.2).abs() < 0.001);
-        assert!((restored_fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_outer_handles_set_crossfade_lengths_without_resizing_selection() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.2, 0.6).with_fade_in(0.25, 0.2));
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInOuterStart,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.1 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade = selection.fade_in().expect("fade-in after outer drag");
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert!((fade.length - 0.25).abs() < 0.001);
-        assert!((fade.mute - 0.25).abs() < 0.001);
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInOuterStart,
-            visible_ratio: 0.1,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.2 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade = selection.fade_in().expect("fade-in should remain");
-        assert!((fade.length - 0.25).abs() < 0.001);
-        assert!(fade.mute.abs() < 0.001);
-
-        state.edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.2, 0.6).with_fade_out(0.25, 0.7));
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutOuterEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.7 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade = selection.fade_out().expect("fade-out after outer drag");
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert!((fade.length - 0.25).abs() < 0.001);
-        assert!((fade.mute - 0.25).abs() < 0.001);
-    }
-
-    #[test]
-    fn primary_press_on_outer_fade_handle_uses_distinct_handle() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.2, 0.6).with_fade_in(0.25, 0.2));
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(40.0, 40.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("outer fade handle interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginEditFade {
-                handle: WaveformEditFadeHandle::FadeInOuterStart,
-                visible_ratio: 0.2
-            }
-        );
-    }
-
-    #[test]
-    fn edit_fade_bottom_handle_resizes_selection_and_keeps_fade_boundary() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.2, 0.6).with_fade_in(0.25, 0.2));
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInStart,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.1 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade = selection.fade_in().expect("fade-in after resize");
-        assert!((selection.start() - 0.1).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert!((selection.start() + selection.width() * fade.length - 0.3).abs() < 0.001);
-        assert!((fade.curve - 0.2).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_keeps_opposite_fade_boundary_stable() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_out(0.25, 0.7),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.8 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        let fade_in_end = selection.start() + selection.width() * fade_in.length;
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        assert!((fade_in_end - 0.3).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_in.curve - 0.2).abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_keeps_crossfade_handles_stable() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_in_mute(0.25)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(0.25),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.7 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        let fade_in_end = selection.start() + selection.width() * fade_in.length;
-        let fade_in_outer_start = selection.start() - selection.width() * fade_in.mute;
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.7).abs() < 0.001);
-        assert!((fade_in_end - 0.3).abs() < 0.001);
-        assert!((fade_in_outer_start - 0.1).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_out_outer_end - 0.7).abs() < 0.001);
-        assert!((fade_in.curve - 0.2).abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_preserves_crossfade_when_fade_collapses() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(1.0),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.5 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_out = selection
-            .fade_out()
-            .expect("fade-out silence handle should remain");
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.start() - 0.2).abs() < 0.001);
-        assert!((selection.end() - 0.5).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_out_outer_end - 1.0).abs() < 0.001);
-        assert!(fade_out.length.abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_does_not_pick_up_silence_during_same_drag() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(1.0),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 1.0 });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.7 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_out = selection
-            .fade_out()
-            .expect("fade-out silence handle should remain");
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.end() - 0.7).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_out_outer_end - 1.0).abs() < 0.000_001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_keeps_collapsed_silence_after_release() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(1.0),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 1.0 });
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 1.0,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.7 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.end() - 0.7).abs() < 0.001);
-        assert!((fade_out_outer_end - 0.7).abs() < 0.000_001);
-    }
-
-    #[test]
-    fn double_click_outer_fade_handles_collapses_silence_without_clearing_fade() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_in_mute(0.5)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(0.75),
-        );
-
-        state.apply_interaction(WaveformInteraction::ClearEditFadeSilence {
-            handle: WaveformEditFadeHandle::FadeInOuterStart,
-        });
-        state.apply_interaction(WaveformInteraction::ClearEditFadeSilence {
-            handle: WaveformEditFadeHandle::FadeOutOuterEnd,
-        });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        assert!((fade_in.length - 0.25).abs() < 0.001);
-        assert!((fade_in.curve - 0.2).abs() < 0.001);
-        assert!(fade_in.mute.abs() < 0.001);
-        assert!((fade_out.length - 0.25).abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-        assert!(fade_out.mute.abs() < 0.001);
-    }
-
-    #[test]
-    fn double_click_on_outer_fade_handle_emits_silence_clear_interaction() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(0.25),
-        );
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerDoubleClick {
-                    position: Point::new(140.0, 40.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("outer fade double-click interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::ClearEditFadeSilence {
-                handle: WaveformEditFadeHandle::FadeOutOuterEnd
-            }
-        );
-    }
-
-    #[test]
-    fn edit_fade_out_top_handle_preserves_silence_after_bottom_handle_collapse() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(1.0),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.5 });
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutStart,
-            visible_ratio: 0.5,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.45,
-        });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_out = selection
-            .fade_out()
-            .expect("fade-out silence handle should remain");
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.end() - 0.5).abs() < 0.001);
-        assert!((fade_out_start - 0.45).abs() < 0.001);
-        assert!((fade_out_outer_end - 1.0).abs() < 0.000_001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_keeps_left_crossfade_pinned_to_sample_edge() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_in_mute(0.5)
-                .with_fade_out(0.25, 0.7),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.7 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_in_outer_start = selection.start() - selection.width() * fade_in.mute;
-
-        assert!(fade_in_outer_start.abs() < 0.000_001);
-    }
-
-    #[test]
-    fn edit_fade_out_bottom_handle_keeps_left_crossfade_pinned_across_wiggles() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_in_mute(0.5)
-                .with_fade_out(0.25, 0.7),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeOutEnd,
-            visible_ratio: 0.6,
-        });
-        for visible_ratio in [0.7, 0.69, 0.71, 0.705, 0.7] {
-            state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio });
-            let selection = state.edit_selection().expect("edit selection");
-            let fade_in = selection.fade_in().expect("fade-in should remain");
-            let fade_in_outer_start = selection.start() - selection.width() * fade_in.mute;
-            assert!(
-                fade_in_outer_start.abs() < 0.000_001,
-                "left silence handle drifted to {fade_in_outer_start}"
-            );
-        }
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.7 });
-    }
-
-    #[test]
-    fn edit_fade_in_bottom_handle_keeps_opposite_fade_boundary_stable() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_out(0.25, 0.7),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInStart,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.1 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        let fade_in_end = selection.start() + selection.width() * fade_in.length;
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        assert!((fade_in_end - 0.3).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_in.curve - 0.2).abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn edit_fade_in_bottom_handle_keeps_crossfade_handles_stable() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.25, 0.2)
-                .with_fade_in_mute(0.25)
-                .with_fade_out(0.25, 0.7)
-                .with_fade_out_mute(0.25),
-        );
-
-        state.apply_interaction(WaveformInteraction::BeginEditFade {
-            handle: WaveformEditFadeHandle::FadeInStart,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection { visible_ratio: 0.1 });
-
-        let selection = state.edit_selection().expect("edit selection");
-        let fade_in = selection.fade_in().expect("fade-in should remain");
-        let fade_out = selection.fade_out().expect("fade-out should remain");
-        let fade_in_end = selection.start() + selection.width() * fade_in.length;
-        let fade_in_outer_start = selection.start() - selection.width() * fade_in.mute;
-        let fade_out_start = selection.end() - selection.width() * fade_out.length;
-        let fade_out_outer_end = selection.end() + selection.width() * fade_out.mute;
-
-        assert!((selection.start() - 0.1).abs() < 0.001);
-        assert!((selection.end() - 0.6).abs() < 0.001);
-        assert!((fade_in_end - 0.3).abs() < 0.001);
-        assert!((fade_in_outer_start - 0.1).abs() < 0.001);
-        assert!((fade_out_start - 0.5).abs() < 0.001);
-        assert!((fade_out_outer_end - 0.7).abs() < 0.001);
-        assert!((fade_in.curve - 0.2).abs() < 0.001);
-        assert!((fade_out.curve - 0.7).abs() < 0.001);
-    }
-
-    #[test]
-    fn primary_press_on_edit_fade_handle_starts_fade_drag_instead_of_playmark() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        let mut widget = waveform_widget_for_state(&state);
-        let bounds = Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0));
-
-        let output = widget
-            .handle_input(
-                bounds,
-                WidgetInput::PointerPress {
-                    position: Point::new(40.0, 4.0),
-                    button: PointerButton::Primary,
-                    modifiers: Default::default(),
-                },
-            )
-            .expect("fade handle interaction");
-        let interaction = output
-            .typed_ref::<WaveformInteraction>()
-            .copied()
-            .expect("waveform interaction");
-
-        assert_eq!(
-            interaction,
-            WaveformInteraction::BeginEditFade {
-                handle: WaveformEditFadeHandle::FadeInEnd,
-                visible_ratio: 0.2
-            }
-        );
-    }
-
-    #[test]
-    fn primary_click_without_drag_still_starts_playback_from_click() {
-        let mut state = WaveformState::synthetic_for_tests();
-
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Play,
-            visible_ratio: 0.45,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.45,
-        });
-
-        assert!(state.is_playing());
-        assert_eq!(state.playhead_ratio(), Some(0.45));
-        assert_eq!(state.play_mark_ratio(), Some(0.45));
-        assert_eq!(state.play_selection(), None);
-    }
-
-    #[test]
-    fn selection_range_projects_visible_ratios_inside_viewport() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Edit,
-            visible_ratio: 0.25,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection {
-            visible_ratio: 0.75,
-        });
-        state.apply_interaction(WaveformInteraction::FinishSelection {
-            visible_ratio: 0.75,
-        });
-        let widget = waveform_widget_for_state(&state);
-        let (start, end) = widget
-            .visible_range_for_selection(state.edit_selection())
-            .expect("selection range");
-
-        assert!((start - 0.25).abs() < 0.001);
-        assert!((end - 0.75).abs() < 0.001);
-    }
-
-    #[test]
-    fn selection_fill_paints_as_overlay_widget_rects() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.apply_interaction(WaveformInteraction::BeginSelection {
-            kind: WaveformSelectionKind::Play,
-            visible_ratio: 0.2,
-        });
-        state.apply_interaction(WaveformInteraction::UpdateSelection { visible_ratio: 0.6 });
-        let widget = waveform_widget_for_state(&state);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        assert!(
-            !primitives
-                .iter()
-                .any(|primitive| matches!(primitive, PaintPrimitive::GpuSurface(_))),
-            "ordinary waveform overlay widget must not emit the GPU waveform"
-        );
-        let fills = fill_rects(&primitives);
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.min.x - 40.0).abs() < 0.001
-                && (fill.rect.max.x - 120.0).abs() < 0.001
-                && (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (255, 142, 92, 48)
-        }));
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x - 40.0).abs() < 1.0
-                && (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (255, 142, 92, 230)
-        }));
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x - 120.0).abs() < 1.0
-                && (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (255, 142, 92, 230)
-        }));
-    }
-
-    #[test]
-    fn edit_selection_paints_start_and_end_boundary_lines() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(wavecrate::selection::SelectionRange::new(0.2, 0.6));
-        let widget = waveform_widget_for_state(&state);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let fills = fill_rects(&primitives);
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x - 40.0).abs() < 1.0
-                && (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (82, 168, 255, 230)
-        }));
-        assert!(fills.iter().any(|fill| {
-            (fill.rect.center().x - 120.0).abs() < 1.0
-                && (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (82, 168, 255, 230)
-        }));
-    }
-
-    #[test]
-    fn edit_fade_curve_paints_volume_trace_as_overlay_rects() {
-        let mut state = WaveformState::synthetic_for_tests();
-        state.edit_selection = Some(
-            wavecrate::selection::SelectionRange::new(0.2, 0.6)
-                .with_fade_in(0.5, 0.8)
-                .with_fade_out(0.25, 0.0),
-        );
-        let widget = waveform_widget_for_state(&state);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let curve_points = fill_rects(&primitives)
-            .into_iter()
-            .filter(|fill| {
-                (fill.color.r, fill.color.g, fill.color.b, fill.color.a) == (82, 168, 255, 225)
-            })
-            .count();
-        assert!(
-            curve_points >= 16,
-            "expected visible fade curve trace points, got {curve_points}"
-        );
-    }
-
-    #[test]
-    fn signal_widget_paints_gpu_surface_without_app_overlay_handles() {
-        let state = WaveformState::synthetic_for_tests();
-        let widget = WaveformSignalWidget::new(
-            state.file(),
-            state.viewport(),
-            state.edit_selection(),
-            state.active_drag_kind(),
-        );
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let surface = primitives
-            .iter()
-            .find_map(|primitive| match primitive {
-                PaintPrimitive::GpuSurface(surface)
-                    if matches!(
-                        surface.content,
-                        GpuSurfaceContent::SignalSummaryBands { .. }
-                    ) =>
-                {
-                    Some(surface)
-                }
-                _ => None,
-            })
-            .expect("waveform gpu surface");
-
-        assert!(surface.overlays.is_empty());
-    }
-
-    #[test]
-    fn signal_widget_attaches_active_edit_fade_gain_preview() {
-        let file = Arc::new(waveform_file_from_mono_samples(
-            "fade-preview.wav".into(),
-            Arc::from([]),
-            48_000,
-            1,
-            vec![1.0; 16],
-        ));
-        let viewport = super::WaveformViewport::full(file.frames);
-        let edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.0, 1.0).with_fade_in(1.0, 0.0));
-        let widget = WaveformSignalWidget::new(Arc::clone(&file), viewport, edit_selection, None);
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let surface = primitives
-            .iter()
-            .find_map(|primitive| match primitive {
-                PaintPrimitive::GpuSurface(surface) => Some(surface),
-                _ => None,
-            })
-            .expect("waveform gpu surface");
-
-        assert!(surface.revision > 0);
-        let GpuSurfaceContent::SignalSummaryBands {
-            summary,
-            gain_preview,
-            ..
-        } = &surface.content
-        else {
-            panic!("expected signal summary bands");
-        };
-        assert!(Arc::ptr_eq(summary, &file.gpu_signal_summary));
-        let preview = gain_preview.expect("edit fade gain preview");
-        assert_eq!(preview.start, 0.0);
-        assert_eq!(preview.end, 1.0);
-        assert_eq!(preview.fade_in_length, 1.0);
-        assert_eq!(preview.fade_in_curve, 0.0);
-    }
-
-    #[test]
-    fn signal_widget_revision_changes_when_same_path_audio_bytes_change() {
-        let first = Arc::new(waveform_file_from_mono_samples(
-            "same-path.wav".into(),
-            Arc::from([1_u8, 2, 3, 4]),
-            48_000,
-            1,
-            vec![0.25; 16],
-        ));
-        let second = Arc::new(waveform_file_from_mono_samples(
-            "same-path.wav".into(),
-            Arc::from([4_u8, 3, 2, 1]),
-            48_000,
-            1,
-            vec![1.0; 16],
-        ));
-
-        let first_revision = gpu_surface_revision_for_file(first);
-        let second_revision = gpu_surface_revision_for_file(second);
-
-        assert_ne!(first_revision, second_revision);
-    }
-
-    #[test]
-    fn signal_widget_keeps_summary_cached_during_live_edit_fade_drag() {
-        let file = Arc::new(waveform_file_from_mono_samples(
-            "fade-preview.wav".into(),
-            Arc::from([]),
-            48_000,
-            1,
-            vec![1.0; 16],
-        ));
-        let viewport = super::WaveformViewport::full(file.frames);
-        let edit_selection =
-            Some(wavecrate::selection::SelectionRange::new(0.0, 1.0).with_fade_in(1.0, 0.0));
-        let widget = WaveformSignalWidget::new(
-            Arc::clone(&file),
-            viewport,
-            edit_selection,
-            Some(WaveformActiveDragKind::EditFade(
-                WaveformEditFadeHandle::FadeInEnd,
-            )),
-        );
-        let mut primitives = Vec::new();
-
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-
-        let surface = primitives
-            .iter()
-            .find_map(|primitive| match primitive {
-                PaintPrimitive::GpuSurface(surface) => Some(surface),
-                _ => None,
-            })
-            .expect("waveform gpu surface");
-
-        assert!(surface.revision > 0);
-        let GpuSurfaceContent::SignalSummaryBands {
-            summary,
-            gain_preview,
-            ..
-        } = &surface.content
-        else {
-            panic!("expected signal summary bands");
-        };
-        assert!(Arc::ptr_eq(summary, &file.gpu_signal_summary));
-        assert!(gain_preview.is_some());
-    }
-
-    fn waveform_widget_for_state(state: &WaveformState) -> WaveformWidget {
-        WaveformWidget::new(WaveformWidgetProps::from_state(state))
-    }
-
-    fn fill_rects(primitives: &[PaintPrimitive]) -> Vec<&PaintFillRect> {
-        primitives
-            .iter()
-            .filter_map(|primitive| match primitive {
-                PaintPrimitive::FillRect(fill) => Some(fill),
-                _ => None,
-            })
-            .collect()
-    }
-
-    fn gpu_surface_revision_for_file(file: Arc<super::WaveformFile>) -> u64 {
-        let viewport = super::WaveformViewport::full(file.frames);
-        let widget = WaveformSignalWidget::new(file, viewport, None, None);
-        let mut primitives = Vec::new();
-        widget.append_paint(
-            &mut primitives,
-            Rect::from_min_size(Point::new(0.0, 0.0), Vector2::new(200.0, 80.0)),
-            &Default::default(),
-            &ThemeTokens::default(),
-        );
-        primitives
-            .iter()
-            .find_map(|primitive| match primitive {
-                PaintPrimitive::GpuSurface(surface) => Some(surface.revision),
-                _ => None,
-            })
-            .expect("waveform gpu surface")
-    }
-
-    fn write_test_wav_i16(path: &std::path::Path, samples: &[i16]) {
-        let spec = hound::WavSpec {
-            channels: 1,
-            sample_rate: 48_000,
-            bits_per_sample: 16,
-            sample_format: hound::SampleFormat::Int,
-        };
-        let mut writer = hound::WavWriter::create(path, spec).expect("create wav");
-        for sample in samples {
-            writer.write_sample(*sample).expect("write sample");
-        }
-        writer.finalize().expect("finalize wav");
-    }
-
-    fn write_test_wav_i16_stereo(path: &std::path::Path, frames: &[(i16, i16)]) {
-        let spec = hound::WavSpec {
-            channels: 2,
-            sample_rate: 48_000,
-            bits_per_sample: 16,
-            sample_format: hound::SampleFormat::Int,
-        };
-        let mut writer = hound::WavWriter::create(path, spec).expect("create wav");
-        for (left, right) in frames {
-            writer.write_sample(*left).expect("write left sample");
-            writer.write_sample(*right).expect("write right sample");
-        }
-        writer.finalize().expect("finalize wav");
-    }
-
-    fn read_test_wav_i16(path: &std::path::Path) -> Vec<i16> {
-        let mut reader = hound::WavReader::open(path).expect("open wav");
-        reader
-            .samples::<i16>()
-            .collect::<Result<Vec<_>, _>>()
-            .expect("read samples")
-    }
-}
+mod tests;
